core specialization fixes
This includes some bug fixes and enhancements to core specialization
commit e0ebb0baca8bb2722cca20dd5c541a839311c2f2

+++ src/plugins/select/bluegene/bg_job_place.c
@@ -1761,7 +1761,7 @@ extern int submit_job(struct job_record *job_ptr, bitstr_t *slurm_block_bitmap,
 	if (!job_ptr->details)
 		return EINVAL;
 
-	if (job_ptr->details->core_spec != NO_VAL) {
+	if (job_ptr->details->core_spec != (uint16_t) NO_VAL) {
 		verbose("select/bluegene: job %u core_spec(%u) not supported",
 			job_ptr->job_id, job_ptr->details->core_spec);
 		job_ptr->details->core_spec = (uint16_t) NO_VAL;
