+++ src/plugins/select/linear/select_linear.c
@@ -2837,7 +2837,7 @@ extern int select_p_job_test(struct job_record *job_ptr, bitstr_t *bitmap,
 		return EINVAL;
 	}
 
-	if (job_ptr->details->core_spec != NO_VAL) {
+	if (job_ptr->details->core_spec != (uint16_t) NO_VAL) {
 		verbose("select/linear: job %u core_spec(%u) not supported",
 			job_ptr->job_id, job_ptr->details->core_spec);
 		job_ptr->details->core_spec = (uint16_t) NO_VAL;
